SystemCode/pip_requirements.txt
pip==21.0.1
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.7
django-model-utils==4.0.0
django-pandas==0.6.1
djangorestframework==3.12.2
joblib==1.0.1
numpy==1.20.0
olefile==0.46
pandas==1.2.3
Pillow==8.1.2
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==1.0.1
scikit-surprise==1.1.1
scipy==1.6.1
setuptools==52.0.0.post20210125
six==1.15.0
sqlparse==0.4.1
threadpoolctl==2.1.0
wheel==0.36.2
wincertstore==0.2
tensorflow==2.7.0
matplotlib==3.4.3
xgboost==1.5.0
transformers==4.11.3
tensorflow_addons==0.14.0
keras==2.6.0
torch==1.10.0+cu113
vaderSentiment==3.3.2
sentence-transformers==2.1.0
ntlk==3.6.5
lightgbm==3.3.1
